fix: 좋아요 API Restful하게 변경

+++ src/main/java/com/seoultech/sanEseo/like/application/service/LikeService.java
@@ -3,6 +3,8 @@ package com.seoultech.sanEseo.like.application.service;
 
 import com.seoultech.sanEseo.like.application.port.LikePort;
 import com.seoultech.sanEseo.like.domain.Likes;
+import com.seoultech.sanEseo.like.exception.DuplicateLikesException;
+import com.seoultech.sanEseo.like.exception.MemberNotLikedException;
 import com.seoultech.sanEseo.member.application.port.out.MemberPort;
 import com.seoultech.sanEseo.member.domain.Member;
 import com.seoultech.sanEseo.post.application.port.PostPort;
@@ -22,15 +24,15 @@ public class LikeService {
     private final MemberPort memberPort;
     private final PostPort postPort;
 
-
-
-    public void addLike(Long memberId, AddLikeRequest request) {
-
-        // request를 Like 객체로 변환
-        Long postId = request.getPostId();
+    public void addLike(Long memberId, Long postId) {
 
         Member member = memberPort.loadById(memberId);
         Post post = postPort.getPost(postId);
+
+        boolean isLiked = likePort.existsByPostAndMember(post, member);
+        if(isLiked) {
+            throw new DuplicateLikesException("이미 좋아요 한 게시글입니다.");
+        }
 
         Likes likes = Likes.builder()
                 .member(member)
@@ -46,15 +48,23 @@ public class LikeService {
         Member member = memberPort.loadById(memberId);
         Post post = postPort.getPost(postId);
 
+        boolean isLiked = likePort.existsByPostAndMember(post, member);
+        if(!isLiked) {
+            throw new MemberNotLikedException("좋아요 하지 않은 게시글입니다.");
+        }
+
         likePort.deleteByPostAndMember(post, member);
     }
 
     public int getLikeCount(Long postId) {
-        return likePort.countByPostId(postId);
+        Post post = postPort.getPost(postId);
+        return likePort.countByPost(post);
     }
 
     public boolean hasMemberLikedPost(Long memberId, Long postId) {
-        return likePort.existsByPostIdAndMemberId(memberId, postId);
+        Member member = memberPort.loadById(memberId);
+        Post post = postPort.getPost(postId);
+        return likePort.existsByPostAndMember(post, member);
     }
 
 }
